reprebot/refac: #5 move vector store crud operations to its own module
- create `crud` submodule inside `vector_store` module
- move crud operations to the new `crud` submodule
